+++ CHANGELOG.md
@@ -9,3 +9,14 @@ Initial version
 #1.1.5
 - Implemented default language, currency and metrics set according to browser locale
 - Implemented settings to configure text selection color
+
+#2.0
+- New extension name and icon
+- Implemented textfield-specific actions (can be disabled from settings)
+- Added ability to remove text selection on tooltip's button click
+- Workarounds for websites that override tooltip's styling 
+- Implemented ability to drag tooltip by the arrow
+- Implemented ability to shift tooltip up when website has it's own selection tooltip (supports Medium and websites with similar tooltip)
+- Added ability to display icons for buttons
+- Implemented pop-up menu with settings
+- Bug fixes and improvements